Refactor server list
